Updated train station and related classes.
-Deleted useless class TrainStationTracker, since we don't need it.
-Moved train listeners to listener package
-Made several variables in the TrainStation class final, because we
don't want them to change. or do we?
-Created getSurface algorithm to fix the station listener, to improve
metrics. Still need to fix, to make it so that the Y variable is omitted
from the check of blocks.
-Did some stuff with the Departure class, and scheduled it to happen,
but it does nothing for now, since we have no data for the database.
-Added some TEMPORARY debug code to CreateTrainStation. until the
Corners Algorithm is fixed, we have no way of knowing which location is
which corner.
-Fixed the bank table problem with mapping player table to the wrong
variable.